+++ docs/CHANGELOG.md
@@ -1,3 +1,12 @@
+## Version ["v0.8.0"] - 2024-12-21
+
+### Features
+- Add config foodHealingValue
+
+### Bug fixes
+- Fix c2me conflict([#93](https://github.com/Okabintaro/UntitledDuckMod/issues/93))
+- Fix ([#94](https://github.com/Okabintaro/UntitledDuckMod/issues/94))
+
 ## Version ["v0.7.0"] - 2024-07-17
 
 ### Features